requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asttokens==2.4.1
    # via stack-data
cardiac-geometries==0.11.0
    # via -r requirements.in
click==8.1.7
    # via rich-click
comm==0.2.0
    # via ipywidgets
decorator==5.1.1
    # via ipython
exceptiongroup==1.1.3
    # via ipython
executing==2.0.1
    # via stack-data
fenics-plotly==2022.0.0
    # via -r requirements.in
ipython==8.17.2
    # via ipywidgets
ipywidgets==8.1.1
    # via fenics-plotly
jedi==0.19.1
    # via ipython
jupyterlab-widgets==3.0.9
    # via ipywidgets
markdown-it-py==3.0.0
    # via rich
matplotlib-inline==0.1.6
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
meshio==5.3.4
    # via cardiac-geometries
numpy==1.26.1
    # via
    #   cardiac-geometries
    #   fenics-plotly
    #   meshio
    #   pandas
    #   pulse2
    #   scipy
packaging==23.2
    # via plotly
pandas==2.1.2
    # via -r requirements.in
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
plotly==5.18.0
    # via fenics-plotly
prompt-toolkit==3.0.41
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pulse2 @ git+https://github.com/finsberg/pulse2.git
    # via -r requirements.in
pure-eval==0.2.2
    # via stack-data
pygments==2.16.1
    # via
    #   ipython
    #   rich
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via pandas
rich==13.6.0
    # via
    #   meshio
    #   rich-click
rich-click==1.7.1
    # via cardiac-geometries
scipy==1.11.3
    # via pulse2
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
stack-data==0.6.3
    # via ipython
tenacity==8.2.3
    # via plotly
tqdm==4.66.1
    # via pulse2
traitlets==5.13.0
    # via
    #   comm
    #   ipython
    #   ipywidgets
    #   matplotlib-inline
typing-extensions==4.8.0
    # via
    #   pulse2
    #   rich-click
tzdata==2023.3
    # via pandas
wcwidth==0.2.10
    # via prompt-toolkit
widgetsnbextension==4.0.9
    # via ipywidgets
